+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-08-31 14:08:38
-// Commit: fix(core): correct layout constraints and overflow in core
+// Updated at: 2026-08-31 16:04:06
+// Commit: feat(core): add bi-directional event bus handler for core
 
 class coreComponentState {
-  final String commitRef = '4075';
-  final String lastMessage = 'fix(core): correct layout constraints and overflow in core';
-  final DateTime timestamp = DateTime.parse('2026-08-31T14:08:38+05:00');
+  final String commitRef = '4079';
+  final String lastMessage = 'feat(core): add bi-directional event bus handler for core';
+  final DateTime timestamp = DateTime.parse('2026-08-31T16:04:06+05:00');
 }